+++ opencompass/tasks/openicl_eval.py
@@ -172,10 +172,7 @@ class OpenICLEvalTask(BaseTask):
                                    if self.output_column else None)
             preds['test_set'] = test_set
             if self.eval_cfg.get('analyze_contamination', False):
-                assert self.dataset_cfg['reader_cfg'].get('contamination_column', None) is not None, (
-                    'You must specify contamination_column in reader_cfg'
-                    'if you want to analyze contamination.')
-                preds['is_cleans'] = test_set[self.contamination_column]
+                preds['is_cleans'] = test_set['is_clean']
             preds = {
                 k: preds[k]
                 for k in signature(icl_evaluator.score).parameters
